game_requirements.txt
pygame==2.5.2
requests==2.31.0
Pillow==10.1.0
